Extracting parent class for preservation_catalog audit
